# web/dashboard/internal/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package db

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Device struct {
	ID        pgtype.UUID
	DeviceEui string
	RwToken   string
	RoToken   string
	Name      string
	Latitude  pgtype.Float8
	Longitude pgtype.Float8
}

type DeviceChannelMapping struct {
	DeviceID        pgtype.UUID
	ChannelID       int16
	Name            pgtype.Text
	MeasurementType pgtype.Int2
	Hidden          bool
}

type DeviceGroup struct {
	DeviceID   pgtype.UUID
	GroupID    pgtype.UUID
	IsReadonly bool
}

type DeviceMeasurement struct {
	ReceivedAt      pgtype.Timestamptz
	DeviceID        pgtype.UUID
	ChannelID       int16
	Value           float64
	MeasurementType int16
}

type Group struct {
	ID      pgtype.UUID
	Name    string
	RwToken string
	RoToken string
}